requirements.txt
pyjwt==2.10.0
bcrypt==4.3.0
requests
python-dotenv
cryptography